requirements.txt
fastapi
uvicorn
boto3
kafka-python
pydantic
scikit-learn